refactor(BA-6482): move scoped_search to non-admin AppConfigPolicyRepository
Scoped search is a non-admin (scope-required) read; admin global search stays
on the admin repository.
